MemberDAO changed Select method to public MemberDTO Select (MemberDTO objDTO);
Removed the select and DALayer_Select from the sqlCrudCommands

Changed the method for select from the interface IDAOCRUDCommands to return generic rather than a bool value.

changed the frmPurchaseOptions text control removed lblLastName and changed lblFirstName to lblName and the text value to
Name

Made the member info such as Name and Points appear on frmPurchases

Changes made in the database side to seperate products into movieticket and food to make it easier to manage data.

All changes are reflected on the documentation